+++ src/com/content-item/item.js
@@ -15,6 +15,8 @@ import ContentCommonNavButton			from 'com/content-common/common-nav-button';
 
 import ContentSimple					from 'com/content-simple/simple';
 
+import VoteResults						from 'com/vote-results/results';
+
 
 import $Node							from '../../shrub/js/node/node';
 import $NodeMeta						from '../../shrub/js/node/node_meta';
@@ -328,18 +330,9 @@ export default class ContentItem extends Component {
 		);
 	}
 
-	render( props, state ) {
-		props = Object.assign({}, props);
+	updatePropsAndGetCategory(node) {
+		let Category = '/';
 		
-		var node = props.node;
-		var user = props.user;
-		var path = props.path;
-		var extra = props.extra;
-		var featured = props.featured;
-		var parent = state.parent;
-		
-		var Category = '/';
-
 		if ( node ) {
 			if ( node.subtype == 'game' ) {
 				props.header = "GAME";
@@ -379,6 +372,21 @@ export default class ContentItem extends Component {
 			
 			props.draft = "Game";
 		}
+		return Category;
+	}
+	
+	render( props, state ) {
+		props = Object.assign({}, props);
+		
+		const node = props.node;
+		const user = props.user;
+		const path = props.path;
+		var extra = props.extra;
+		var featured = props.featured;
+		const parent = state.parent;
+		
+		// This seems wrong, it would make more sense to update state.
+		const Category = updatePropsAndGetCategory(node);
 		
 		var ShowEventPicker = null;
 		if ( extra && extra.length && extra[0] == 'edit' && node_CanPublish(parent) ) {
@@ -590,50 +598,10 @@ export default class ContentItem extends Component {
 			}
 		}
 		// Final Results
-		else if ( !parseInt(node_CanGrade(parent)) && node_isEventFinished(parent) ) {
+		else {
 			// grading is closed
-			let Lines = [];
+			ShowGrade = (<VoteResults />);
 
-			for ( var key in parent.meta ) {
-				// Is it a valid grade ?
-				let parts = key.split('-');
-				if ( parts.length == 2 && parts[0] == 'grade' ) {
-					// Make sure they user hasn't opted out
-					
-					if ( node.meta && !(node.meta[key+'-out']|0) ) {
-						Lines.push({'key': key, 'value': parent.meta[key]});
-					}
-				}
-			}
-			
-			let ResultLines = [];
-			for ( let idx = 0; idx < Lines.length; idx++ ) {
-				let Line = Lines[idx];
-				
-				let Title = Line.value;
-				let Place = "N/A";
-				if ( node.magic && node.magic[Line.key+'-result'] )
-					Place = node.magic[Line.key+'-result'];
-				let Average = 0;
-				if ( node.magic && node.magic[Line.key+'-average'] )
-					Average = node.magic[Line.key+'-average'];
-				let Count = 0;
-				if ( node.grade && node.grade[Line.key] )
-					Count = node.grade[Line.key];
-				
-				//  {Score >= 20 ? <SVGIcon small baseline>check</SVGIcon> : <SVGIcon small baseline>cross</SVGIcon>}
-				
-				ResultLines.push(<div class="-grade"><span class="-title">{Title}:</span> <strong>{Place}</strong><sup>{this.positionSuffix(Place)}</sup> ({Average} average from {Count} ratings)</div>);
-			}
-
-			ShowGrade = (
-				<ContentCommonBody class="-rating">
-					<div class="-header">Results</div>
-					<div class="-subtext">Final results</div>
-					<div class="-items">{ResultLines}</div>
-					<div class="-footer">When a line is <strong>N/A</strong>, it means there weren't enough ratings for a reliable score. Don't forget to play and rate other people's games during events to prioritize your game.</div>
-				</ContentCommonBody>
-			); //'
 		}
 		
 		var ShowOptOut = null;
